Started scene
A scene is just a collection of physics items. you can make all the items in the scene move with either of the move functions. You can add items to a scene. And if you already know the indeces of two items in the scene, you can check them for collisions to one another