began visualizer
we now have a frame (window) that holds a panel (graphics). The panel has a scene, and the scene is the collection of physics items. each item is drawn, and a move call from the frame level is passed to the item level to move the items in the visual scene