Began ship shooting function
Shooting function started. Currently lives in HellGame class, but should probably exist in a ship class (so we can make modifications to the ship, like upgrades to the weapon, etc). Need to implement cooldown timer, to  make gaps between bullets. Also added a little more art to the images folder. The objectively best one is the snail wizard, which is inarguably the greatest creation of human hands in history.